frontend: App: project: Resolve selectors for project views
Avoid empty project queries and preserve loaded results with no namespaces.
