+++ frontend/src/components/project/ProjectList.tsx
@@ -24,6 +24,7 @@ import Namespace from '../../lib/k8s/namespace';
 import { HeadlampEventType, useEventCallback } from '../../redux/headlampEventSlice';
 import { useTypedSelector } from '../../redux/hooks';
 import { ProjectDefinition } from '../../redux/projectsSlice';
+import AllowedNamespacesSelectorGate from '../App/AllowedNamespacesSelectorGate';
 import { StatusLabel } from '../common';
 import Link from '../common/Link';
 import Table, { TableColumn } from '../common/Table/Table';
@@ -77,14 +78,18 @@ export const useProject = (name: string) => {
     () => ({
       isLoading,
       project: namespaces
-        ? groupNamespacesIntoProjects(namespaces).find(project => project.id === name)
+        ? groupNamespacesIntoProjects(namespaces).find(project => project.id === name) ?? {
+            id: name,
+            clusters: [],
+            namespaces: [],
+          }
         : undefined,
     }),
     [namespaces, name, isLoading]
   );
 };
 
-export default function ProjectList() {
+function ProjectListContent() {
   const { t } = useTranslation();
   const [showCreate, setShowCreate] = useState(false);
   const pluginApiResources = useTypedSelector(state => state.projects.apiResources);
@@ -227,5 +232,21 @@ export default function ProjectList() {
 
       <Table key={pluginApiResources.length} columns={columns} data={projects} />
     </>
+  );
+}
+
+/**
+ * Resolves configured namespace selectors before querying the project list.
+ *
+ * @returns The gated project list.
+ */
+export default function ProjectList() {
+  const clusterConf = useClustersConf();
+  const clusters = Object.values(clusterConf ?? {}).map(cluster => cluster.name);
+
+  return (
+    <AllowedNamespacesSelectorGate clusters={clusters}>
+      <ProjectListContent />
+    </AllowedNamespacesSelectorGate>
   );
 }
